pyproject.toml
[project]
name = "receipt-vault"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "asgi-correlation-id>=4.3.4",
    "dramatiq[redis,watch]>=1.18.0",
    "fastapi[standard]>=0.115.14",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pydantic[standard]>=2.11.7",
    "python-magic>=0.4.27",
    "requests>=2.32.4",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.35.0",
]
